% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gardasil.R
\docType{data}
\name{gardasil}
\alias{gardasil}
\title{The Gardasil dataset in R data.frame format}
\format{A data.frame with 1413 rows and 11 variables.
\describe{
  \item{Age}{Age in years}
  \item{AgeGroup}{Categorical age: 18-26 or 17-27}
  \item{Race}{white/black/hispanic/"other/unknown"}
  \item{Shots}{Number of shots completed (1, 2, or 3)}
  \item{Completed}{yes/no, "yes" if all 3 shots in the series were completed}
  \item{InsuranceType}{medical assistance / private payer / hospital based / military}
  \item{MedAssist}{yes/no}
  \item{Location}{Four locations: Odenton, White Marsh, Johns Hopkins Outpatient Center, Bayview}
  \item{LocationType}{urban/suburban}
  \item{PracticeType}{pediatric / family practice / OB-GYN}
  \item{RaceSummary}{white/minority/"other/unknown"}
}}
\source{
Chou B, Krill LS, Horton BB, Barat CE, Trimble CL:
Disparities in human papillomavirus vaccine completion among
vaccine initiators. Obstet. Gynecol. 2011, 118:14–20.
}
\usage{
gardasil
}
\description{
The Gardasil dataset in R data.frame format
}
\details{
Note that a cleaned dataset like this is provided, it is recommended
to place the code that produces it in data-raw. You can then use
usethis::use_data_raw() to update the cleaned dataset in data/. See
http://r-pkgs.had.co.nz/data.html.
}
\keyword{datasets}
