% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gardasil.R
\name{read_gardasil}
\alias{read_gardasil}
\title{Title Read Gardasil data from csv file}
\usage{
read_gardasil(other.as.NA = FALSE)
}
\arguments{
\item{other.as.NA}{If FALSE (default), unknown/other race is treated as its own category.
If TRUE, treat unknown/other race as NA.}
}
\value{
Cleaned/formatted data.frame. This data.frame could be documented here as its
codebook, but such documentation is demonstrated under the prepared
`gardasil` dataset (see `?gardasil`)
}
\description{
Title Read Gardasil data from csv file
}
\details{
The other.as.NA option is provided because the
original manuscript was unclear about how other/unknown race was treated.
Reproducing the reported results requires this parameter to be FALSE, but
its coefficients are not reported.
}
\examples{
gardasil2 <- read_gardasil()
summary(gardasil2)
}
